Broken spreadsheet view
Our users complained about the preview for spreadsheets, and in my research, I discovered that for the preview, someone has chosen to use the embedded version, which didn't work ok with Share.
Besides that, as the Alfresco Share is not a mobile friendly UI, I can not agree with the decision of not using the desktiop version, which I have changed on this commit.